+++ modules/external-link-intercepter/scripts/init.js
@@ -1,0 +1,12 @@
+(function (angular) {
+    'use strict';
+
+    angular.module('externalLinkInterceptor', [
+        'angular-external-link-interceptor.controller',
+        'angular-external-link-interceptor.anchor',
+        'angular-external-link-interceptor.external-link-config',
+        'angular-external-link-interceptor.external-link-service',
+        'angular-external-link-intercepter.templates'
+    ]);
+
+}(window.angular));